Moved RegexFlags class into org.evomaster.core.utils package
